move open to virtual document
