+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/common/storage/VirtualDocument.scala
@@ -1,7 +1,27 @@
 package edu.uci.ics.amber.engine.common.storage
 
+import edu.uci.ics.texera.workflow.common.storage.FileResolver.DATASET_FILE_URI_SCHEME
+
 import java.io.{File, InputStream}
 import java.net.URI
+
+object VirtualDocument {
+  type FileHandle = ReadonlyVirtualDocument[_]
+
+  def openFile(fileUri: URI): FileHandle = {
+    fileUri.getScheme match {
+      case DATASET_FILE_URI_SCHEME =>
+        new DatasetFileDocument(fileUri)
+
+      case "file" =>
+        // For local files, create a ReadonlyLocalFileDocument
+        new ReadonlyLocalFileDocument(fileUri)
+
+      case _ =>
+        throw new UnsupportedOperationException(s"Unsupported URI scheme: ${fileUri.getScheme}")
+    }
+  }
+}
 
 /**
   * TODO: break this base definition into more self-contained pieces, including Writeonly, IteratorBased
