+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/source/scan/csv/CSVScanSourceOpExec.scala
@@ -5,8 +5,7 @@ import edu.uci.ics.amber.engine.common.executor.SourceOperatorExecutor
 import edu.uci.ics.amber.engine.common.workflow.PortIdentity
 import edu.uci.ics.amber.engine.common.{CheckpointState, CheckpointSupport}
 import edu.uci.ics.amber.engine.common.model.tuple.{AttributeTypeUtils, Schema, TupleLike}
-import edu.uci.ics.amber.engine.common.storage.DatasetFileDocument
-import edu.uci.ics.texera.workflow.common.storage.{FileOpener, FileResolver}
+import edu.uci.ics.amber.engine.common.storage.VirtualDocument.openFile
 import edu.uci.ics.texera.workflow.operators.source.scan.FileDecodingMethod
 
 import java.io.InputStreamReader
@@ -70,7 +69,7 @@ class CSVScanSourceOpExec private[csv] (
 
   override def open(): Unit = {
     inputReader = new InputStreamReader(
-      FileOpener.openFile(new URI(fileUri)).asInputStream(),
+      openFile(new URI(fileUri)).asInputStream(),
       fileEncoding.getCharset
     )
