+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/source/scan/json/JSONLScanSourceOpExec.scala
@@ -1,11 +1,10 @@
 package edu.uci.ics.texera.workflow.operators.source.scan.json
 
 import edu.uci.ics.amber.engine.common.executor.SourceOperatorExecutor
-import edu.uci.ics.amber.engine.common.storage.DatasetFileDocument
 import edu.uci.ics.amber.engine.common.Utils.objectMapper
 import edu.uci.ics.amber.engine.common.model.tuple.AttributeTypeUtils.parseField
 import edu.uci.ics.amber.engine.common.model.tuple.{Schema, TupleLike}
-import edu.uci.ics.texera.workflow.common.storage.{FileOpener, FileResolver}
+import edu.uci.ics.amber.engine.common.storage.VirtualDocument.openFile
 import edu.uci.ics.texera.workflow.operators.source.scan.FileDecodingMethod
 import edu.uci.ics.texera.workflow.operators.source.scan.json.JSONUtil.JSONToMap
 
@@ -44,7 +43,7 @@ class JSONLScanSourceOpExec private[json] (
     schema = schemaFunc()
     reader = new BufferedReader(
       new InputStreamReader(
-        FileOpener.openFile(new URI(fileUri)).asInputStream(),
+        openFile(new URI(fileUri)).asInputStream(),
         fileEncoding.getCharset
       )
     )
